refactor dpm campaign launch routes
